completed one part of authentication

+++ base/views.py
@@ -17,6 +17,10 @@ from .forms import RoomForm
 # ]
 
 def loginPage(request):
+
+    if request.user.is_authenticated:
+        return redirect('home')
+
     if request.method == 'POST':
         username = request.POST.get('username')
         password = request.POST.get('password')
@@ -100,6 +104,10 @@ def updateRoom(request, pk):
 @login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('You are not authorized to edit this room')
+        
     if request.method == 'POST':
         room.delete()
         return redirect('home')
